[refactor] ssd 에서 사용하는 파일들을 ssd_modules directory로 이동
