+++ ssd_modules/command_buffer/tests/test_buffer_driver.py
@@ -1,5 +1,5 @@
 import pytest
-from buffer_driver import BufferDriver
+from ssd_modules.command_buffer.buffer_driver import BufferDriver
 import os
 
 @pytest.fixture
@@ -8,7 +8,7 @@ def buffer_driver():
 
 
 def test_make_empty_list(buffer_driver):
-    assert  os.path.isfile(os.path.join('buffer_folder','1_empty'))
+    assert  os.path.isfile(os.path.join('buffer_folder', '1_empty'))
 
 def test_get_list_from_buffer_empty_files(buffer_driver):
     list = []
@@ -41,4 +41,4 @@ def test_make_list_to_buffer_file_fail(buffer_driver):
     assert buffer_driver.make_buffer_files_from_list(list) == None
 def test_delete_buffer_files(buffer_driver):
     buffer_driver.delete_buffer_files()
-    assert not os.path.isfile(os.path.join('buffer_folder','1_empty'))
+    assert not os.path.isfile(os.path.join('../../../buffer_folder', '1_empty'))
