+++ ssd_modules/command_buffer/tests/test_command_buffer_real.py
@@ -1,9 +1,8 @@
 import pytest
-from command_buffer import CommandBuffer
-from buffer_driver import BufferDriver
-from nand import Nand
+from ssd_modules.command_buffer.command_buffer import CommandBuffer
+from ssd_modules.device.nand import Nand
 
-FILE_PATH = "test_ssd_nand.txt"
+FILE_PATH = "../../../test_ssd_nand.txt"
 LBA_LENGTH = 100
 DEFAULT_VALUE = "0x00000000"
 FIRST_ADDRESS = 0
