+++ ssd_modules/commands/write_command.py
